+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## 1.1
 
+- FIX : Page reload prevent action param in url or post - *19/01/2021* - 1.1.3
 - FIX : Missing langs and help infos - *13/12/2021* - 1.1.2
 - FIX : By default, no supplier price preselection on supplier proposal, and conf ADVANCED_PRODUCT_SEARCH_PRESELECT_IF_ONE_FOURN_PRICE_ON_SUPPLIER_PROPOSAL to restore preselection - *08/12/2021* - 1.1.1
 - NEW : handle supplier documents - *08/12/2021* - 1.1.0